+++ Gruntfile.js
@@ -11,6 +11,7 @@ module.exports = function(grunt) {
 	grunt.loadNpmTasks('grunt-html-build');
 	grunt.loadNpmTasks('grunt-grunticon');
 	grunt.loadNpmTasks('grunt-svgmin');
+	grunt.loadNpmTasks('grunt-contrib-connect');
 
 	// config
 	grunt.initConfig({
@@ -63,7 +64,8 @@ module.exports = function(grunt) {
 		sass: {
 			prod: {
 				options: {
-					style: 'compressed'
+					style: 'compressed',
+					noCache: true
 				},
 
 				files: {
@@ -116,7 +118,7 @@ module.exports = function(grunt) {
 				files: [
 					{
 						expand: true,
-						cwd: 'src/assets/img',
+						cwd: 'src/assets/images',
 						src: '{,*/}*.svg',
 						dest: 'dist/images/'
 					},
@@ -161,6 +163,18 @@ module.exports = function(grunt) {
 					dest: 'dist/css/icons/'
 				}
 			}
+		},
+
+		// Connect plugin for server and synchronisation between browsers/devices
+		connect: {
+			server: {
+				options: {
+					port: 9001,
+					hostname: '0.0.0.0',
+					base: 'dist',
+					keepalive: true
+				}
+			}
 		}
 	});
 
@@ -169,4 +183,6 @@ module.exports = function(grunt) {
 	grunt.registerTask('setup', ['sass','svgmin','jshint','uglify','htmlbuild','grunticon']);
 
 	grunt.registerTask('img', ['svgmin','grunticon','svg2png','imageoptim']);
+
+	grunt.registerTask('server', ['connect:server']);
 };